requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.5
google-cloud-videointelligence==2.11.4
google-cloud-storage==2.13.0
moviepy==1.0.3
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
python-instagram==1.3.2
tiktok-api==0.1.0
pydantic==2.5.2
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiohttp==3.9.1
python-dateutil==2.8.2
